test: split the default msw handler monolith into modules
config/test/msw/handlers.ts held 470 lines of inline defaults that every
jest test inherits through the global server. Split into
defaults/gateway.ts, defaults/chains.ts, and defaults/hypernative.ts
with byte-identical response bodies; handlers.ts becomes the
composition and keeps its export, so both consumers
(apps/web and apps/mobile server.ts) are untouched. Verified by the
full web suite: 753 suites, 6801 tests, unchanged results.
